8341413: Stop including osThread_os.hpp in the middle of the OSThread class
