+++ src/hotspot/os/bsd/osThread_bsd.hpp
@@ -25,25 +25,21 @@
 #ifndef OS_BSD_OSTHREAD_BSD_HPP
 #define OS_BSD_OSTHREAD_BSD_HPP
 
- private:
-  int _thread_type;
+#include "runtime/osThreadBase.hpp"
+#include "suspendResume_posix.hpp"
+#include "utilities/globalDefinitions.hpp"
 
- public:
-
-  int thread_type() const {
-    return _thread_type;
-  }
-  void set_thread_type(int type) {
-    _thread_type = type;
-  }
-
- private:
+class OSThread : public OSThreadBase {
+  friend class VMStructs;
 
 #ifdef __APPLE__
   typedef thread_t thread_id_t;
 #else
   typedef pid_t thread_id_t;
 #endif
+
+  thread_id_t _thread_id;
+  int _thread_type;
 
   // _pthread_id is the pthread id, which is used by library calls
   // (e.g. pthread_kill).
@@ -57,15 +53,26 @@
   sigset_t _caller_sigmask; // Caller's signal mask
 
  public:
+  OSThread();
+  ~OSThread();
+
+  int thread_type() const {
+    return _thread_type;
+  }
+  void set_thread_type(int type) {
+    _thread_type = type;
+  }
 
   // Methods to save/restore caller's signal mask
   sigset_t  caller_sigmask() const       { return _caller_sigmask; }
   void    set_caller_sigmask(sigset_t sigmask)  { _caller_sigmask = sigmask; }
 
-#ifndef PRODUCT
-  // Used for debugging, return a unique integer for each thread.
-  intptr_t thread_identifier() const   { return (intptr_t)_pthread_id; }
-#endif
+  thread_id_t thread_id() const {
+    return _thread_id;
+  }
+  void set_thread_id(thread_id_t id) {
+    _thread_id = id;
+  }
 
   pthread_t pthread_id() const {
     return _pthread_id;
@@ -80,7 +87,6 @@
   // suspension support.
   // ***************************************************************
 
-public:
   // flags that support signal based suspend/resume on Bsd are in a
   // separate class to avoid confusion with many flags in OSThread that
   // are used by VM level suspend/resume.
@@ -126,17 +132,9 @@ public:
     return _startThread_lock;
   }
 
-  // ***************************************************************
-  // Platform dependent initialization and cleanup
-  // ***************************************************************
-
-private:
-
-  void pd_initialize();
-  void pd_destroy();
-
-// Reconciliation History
-// osThread_solaris.hpp 1.24 99/08/27 13:11:54
-// End
+  uintx thread_id_for_printing() const override {
+    return (uintx)_thread_id;
+  }
+};
 
 #endif // OS_BSD_OSTHREAD_BSD_HPP
