+++ src/hotspot/os/windows/osThread_windows.cpp
@@ -22,17 +22,17 @@
  *
  */
 
-// no precompiled headers
-#include "runtime/os.hpp"
+#include "precompiled.hpp"
 #include "runtime/osThread.hpp"
 
-void OSThread::pd_initialize() {
-  set_thread_handle(nullptr);
-  set_thread_id(0);
-  set_interrupt_event(nullptr);
-}
+#include <Windows.h>
 
-void OSThread::pd_destroy() {
+OSThread::OSThread()
+  : _thread_id(0),
+    _thread_handle(nullptr),
+    _interrupt_event(nullptr) {}
+
+OSThread::~OSThread() {
   if (_interrupt_event != nullptr) {
     CloseHandle(_interrupt_event);
   }
